+++ src/components/FooterNews.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import svg from '../assets/sprite.svg';
 
 interface IProps {