+++ src/components/Modal.tsx
@@ -1,4 +1,4 @@
-import { useRef } from "react";
+import React, { useRef } from "react";
 
 interface IModal {
     setModal: (e: boolean) => void;