+++ src/components/Widgets/Wallets.tsx
@@ -1,3 +1,9 @@
+import React from 'react';
+import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
+import { Doughnut } from 'react-chartjs-2';
+
+ChartJS.register(ArcElement, Tooltip, Legend);
+
 export default function Wallets() {
     const WidgetToolBar = () => {
         return <div className="widget__toolbar">
@@ -19,9 +25,33 @@ export default function Wallets() {
     }
 
     const WidgetBody = () => {
+        const data = {
+            datasets: [{
+                label: '# of Votes',
+                data: [12, 19, 3, 5, 2, 3],
+                backgroundColor: [
+                    'rgba(255, 99, 132, 0.2)',
+                    'rgba(54, 162, 235, 0.2)',
+                    'rgba(255, 206, 86, 0.2)',
+                    'rgba(75, 192, 192, 0.2)',
+                    'rgba(153, 102, 255, 0.2)',
+                    'rgba(255, 159, 64, 0.2)'
+                ],
+                borderColor: [
+                    'rgba(255, 99, 132, 1)',
+                    'rgba(54, 162, 235, 1)',
+                    'rgba(255, 206, 86, 1)',
+                    'rgba(75, 192, 192, 1)',
+                    'rgba(153, 102, 255, 1)',
+                    'rgba(255, 159, 64, 1)'
+                ],
+                borderWidth: 1
+            }]
+        };
+
         return <div className="items">
             <div className="chart chart__container chart__container--dark">
-                <canvas id="myChart"></canvas>
+                <Doughnut data={data} />
                 <p className="chart__container--subtitle">exchanges dominance</p>
                 <div className="select select__container">
                     <select className="widget__select widget__select--wide" name="" id="">