+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.0.4
+- Added chart.js repository
+- Added lightweight-charts repository
+- Added Chart widget
+- Fixed minor issues
+
 ## v0.0.3
 - Added Modal
 - Added the option to add and remove tabs in widgets